+++ test/test_properties.py
@@ -1,16 +1,28 @@
 from __future__ import annotations
-from collections import namedtuple
+from typing import NamedTuple
 import pytest
 from permutation import Permutation
 
-PermData = namedtuple(
-    "PermData",
-    "p degree order even sign bool mod_lehmer inversions cycles str image image6 permuted",
-)
+
+class PermData(NamedTuple):
+    p: Permutation
+    degree: int
+    order: int
+    even: bool
+    sign: int
+    bool: bool
+    mod_lehmer: int
+    inversions: int
+    cycles: list[tuple[int, ...]]
+    str: str
+    image: tuple[int, ...]
+    image6: tuple[int, ...]
+    permuted: list[int]
+
 
 PERMUTATIONS = [
     PermData(
-        Permutation(), 0, 1, True, 1, False, 0, 0, [], "1", (), (1, 2, 3, 4, 5, 6), ()
+        Permutation(), 0, 1, True, 1, False, 0, 0, [], "1", (), (1, 2, 3, 4, 5, 6), []
     ),
     PermData(
         Permutation(2, 1),
@@ -25,7 +37,7 @@ PERMUTATIONS = [
         "(1 2)",
         (2, 1),
         (2, 1, 3, 4, 5, 6),
-        (2, 1),
+        [2, 1],
     ),
     PermData(
         Permutation(1, 3, 2),
@@ -40,7 +52,7 @@ PERMUTATIONS = [
         "(2 3)",
         (1, 3, 2),
         (1, 3, 2, 4, 5, 6),
-        (1, 3, 2),
+        [1, 3, 2],
     ),
     PermData(
         Permutation(3, 1, 2),
@@ -55,7 +67,7 @@ PERMUTATIONS = [
         "(1 3 2)",
         (3, 1, 2),
         (3, 1, 2, 4, 5, 6),
-        (2, 3, 1),
+        [2, 3, 1],
     ),
     PermData(
         Permutation(2, 3, 1),
@@ -70,7 +82,7 @@ PERMUTATIONS = [
         "(1 2 3)",
         (2, 3, 1),
         (2, 3, 1, 4, 5, 6),
-        (3, 1, 2),
+        [3, 1, 2],
     ),
     PermData(
         Permutation(3, 2, 1),
@@ -85,7 +97,7 @@ PERMUTATIONS = [
         "(1 3)",
         (3, 2, 1),
         (3, 2, 1, 4, 5, 6),
-        (3, 2, 1),
+        [3, 2, 1],
     ),
     PermData(
         Permutation(1, 2, 4, 3),
@@ -100,7 +112,7 @@ PERMUTATIONS = [
         "(3 4)",
         (1, 2, 4, 3),
         (1, 2, 4, 3, 5, 6),
-        (1, 2, 4, 3),
+        [1, 2, 4, 3],
     ),
     PermData(
         Permutation(2, 1, 4, 3),
@@ -115,7 +127,7 @@ PERMUTATIONS = [
         "(1 2)(3 4)",
         (2, 1, 4, 3),
         (2, 1, 4, 3, 5, 6),
-        (2, 1, 4, 3),
+        [2, 1, 4, 3],
     ),
     PermData(
         Permutation(2, 3, 4, 1),
@@ -130,7 +142,7 @@ PERMUTATIONS = [
         "(1 2 3 4)",
         (2, 3, 4, 1),
         (2, 3, 4, 1, 5, 6),
-        (4, 1, 2, 3),
+        [4, 1, 2, 3],
     ),
     PermData(
         Permutation(2, 3, 1, 5, 4),
@@ -145,7 +157,7 @@ PERMUTATIONS = [
         "(1 2 3)(4 5)",
         (2, 3, 1, 5, 4),
         (2, 3, 1, 5, 4, 6),
-        (3, 1, 2, 5, 4),
+        [3, 1, 2, 5, 4],
     ),
     PermData(
         Permutation(2, 3, 4, 5, 1),
@@ -160,7 +172,7 @@ PERMUTATIONS = [
         "(1 2 3 4 5)",
         (2, 3, 4, 5, 1),
         (2, 3, 4, 5, 1, 6),
-        (5, 1, 2, 3, 4),
+        [5, 1, 2, 3, 4],
     ),
     PermData(
         Permutation(5, 4, 3, 2, 1),
@@ -175,7 +187,7 @@ PERMUTATIONS = [
         "(1 5)(2 4)",
         (5, 4, 3, 2, 1),
         (5, 4, 3, 2, 1, 6),
-        (5, 4, 3, 2, 1),
+        [5, 4, 3, 2, 1],
     ),
     PermData(
         Permutation(2, 1, 4, 3, 6, 5),
@@ -190,7 +202,7 @@ PERMUTATIONS = [
         "(1 2)(3 4)(5 6)",
         (2, 1, 4, 3, 6, 5),
         (2, 1, 4, 3, 6, 5),
-        (2, 1, 4, 3, 6, 5),
+        [2, 1, 4, 3, 6, 5],
     ),
     PermData(
         Permutation(2, 3, 4, 1, 6, 5),
@@ -205,7 +217,7 @@ PERMUTATIONS = [
         "(1 2 3 4)(5 6)",
         (2, 3, 4, 1, 6, 5),
         (2, 3, 4, 1, 6, 5),
-        (4, 1, 2, 3, 6, 5),
+        [4, 1, 2, 3, 6, 5],
     ),
     PermData(
         Permutation(2, 3, 1, 5, 6, 4),
@@ -220,7 +232,7 @@ PERMUTATIONS = [
         "(1 2 3)(4 5 6)",
         (2, 3, 1, 5, 6, 4),
         (2, 3, 1, 5, 6, 4),
-        (3, 1, 2, 6, 4, 5),
+        [3, 1, 2, 6, 4, 5],
     ),
     PermData(
         Permutation(2, 3, 4, 5, 6, 1),
@@ -235,7 +247,7 @@ PERMUTATIONS = [
         "(1 2 3 4 5 6)",
         (2, 3, 4, 5, 6, 1),
         (2, 3, 4, 5, 6, 1),
-        (6, 1, 2, 3, 4, 5),
+        [6, 1, 2, 3, 4, 5],
     ),
 ]
 
@@ -332,7 +344,7 @@ def test_from_image(p: Permutation, image: tuple[int, ...]) -> None:
 
 
 @pytest.mark.parametrize("p,permuted", [(d.p, d.permuted) for d in PERMUTATIONS])
-def test_permute(p: Permutation, permuted: tuple[int, ...]) -> None:
+def test_permute(p: Permutation, permuted: list[int]) -> None:
     assert p.permute(range(1, p.degree + 1)) == permuted