+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.5.0 (in development)
 -----------------------
 - Support Python 3.11
+- The `permute()` method now accepts iterables of any element type and returns
+  a list
 
 v0.4.0 (2022-10-21)
 -------------------